Did performance tests on syringe pump and revealed equation relating motor step to volume.

+++ Modules/1008 SYR/ARCaDIUS_mkii_SYR/Syringe_Motor.h
@@ -15,6 +15,6 @@ class StepperMotor
     //Constructor
     StepperMotor() = default;
     void setUp();
-    void pumpVolume(uint16_t volume, uint16_t direction);
+    void pumpVolume(float volume, uint16_t direction);
     void move();
 };
